Move prefix sumtree into osmoutils
