+++ x/lockup/keeper/lock.go
@@ -10,7 +10,7 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/gogo/protobuf/proto"
 
-	"github.com/osmosis-labs/osmosis/v13/store"
+	"github.com/osmosis-labs/osmosis/v13/osmoutils/sumtree"
 	"github.com/osmosis-labs/osmosis/v13/x/lockup/types"
 )
 
@@ -613,8 +613,8 @@ func (k Keeper) SlashTokensFromLockByID(ctx sdk.Context, lockID uint64, coins sd
 	return lock, nil
 }
 
-func (k Keeper) accumulationStore(ctx sdk.Context, denom string) store.Tree {
-	return store.NewTree(prefix.NewStore(ctx.KVStore(k.storeKey), accumulationStorePrefix(denom)), 10)
+func (k Keeper) accumulationStore(ctx sdk.Context, denom string) sumtree.Tree {
+	return sumtree.NewTree(prefix.NewStore(ctx.KVStore(k.storeKey), accumulationStorePrefix(denom)), 10)
 }
 
 // removeTokensFromLock is called by lockup slash function.
